requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
torch==2.2.2
torchvision==0.17.2
Pillow==10.3.0
numpy==1.26.4
matplotlib==3.8.4
boto3==1.34.0